create

+++ main-server/httpClient/src/main/java/ru/practicum/admin/dao/UserRepository.java
@@ -1,12 +1,11 @@
 package ru.practicum.admin.dao;
 
 import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.data.jpa.repository.Query;
-import ru.practicum.admin.model.User;
+import ru.practicum.admin.models.user.User;
 
 import java.util.List;
 
 
-public interface UserRepository extends JpaRepository<User, Long> {
+public interface UserRepository extends JpaRepository<User, Integer> {
     List<User> findByIdIn(List<Integer> ids);
 }
